+++ front-end/js/pages/favoritos.js
@@ -1,0 +1,36 @@
+/* *********************************************************************
+* Objetivo: Template e a lógica de favoritos
+* Data: 09/12/2025
+* Autor: Marcelo Vieira
+* Versão: 1.0
+* **********************************************************************/
+
+export const Favoritos = {
+    title: "FAVORITOS",
+    template: `
+            <div class="favoritos-container">
+                <section class="cards">
+                    <div class="card">
+                        <img src="./img/pet-teste.jpg" alt="Pet">
+                        <h3>Rex</h3>
+                        <p>Disponível</p>
+                    </div>
+                    <div class="card">
+                        <img src="./img/pet-teste.jpg" alt="Pet">
+                        <h3>Bolinha</h3>
+                        <p>Disponível</p>
+                    </div>
+                    <div class="card">
+                        <img src="./img/pet-teste.jpg" alt="Pet">
+                        <h3>Thor</h3>
+                        <p>Disponível</p>
+                    </div>
+                     <div class="card">
+                        <img src="./img/pet-teste.jpg" alt="Pet">
+                        <h3>Jade</h3>
+                        <p>Disponível</p>
+                    </div>
+                </section>
+            </div>
+        `
+};
